surveillance/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
numpy<2
opencv-python==4.9.0.80
aiofiles==23.2.1
